refactor: Move files to src
